refactoring of the serializers
